Update for homeassistant temp setting
is_deviated was previously using abs() because we were trying to be cooling and heating agnostic.
However this became a problem when setting the target temperature because if you set it more than
a degree below the current temperature it would then heat off into infinity.

Broke the function into deviated heating with the intent of making a cooling one in the future